improve board inspecting
